Node e2e documentations and minor features - Add README.md for node e2e tests - Add support for --cleanup=false to leave test files on remote hosts and temporary instances for debugging - Add ubuntu trusty instances for docker 1.8 and docker 1.9 to jenkins pr builder - Disable coreos-beta for jenkins ci since it is failing - need to investigate

+++ test/e2e_node/runner/run_e2e.go
@@ -15,9 +15,9 @@ limitations under the License.
 */
 
 // To run the e2e tests against one or more hosts on gce:
-// $ go run run_e2e.go --logtostderr --v 2 --ssh-env gce --hosts <comma separated hosts>
+// $ godep go run run_e2e.go --logtostderr --v 2 --ssh-env gce --hosts <comma separated hosts>
 // To run the e2e tests against one or more images on gce and provision them:
-// $ go run run_e2e.go --logtostderr --v 2 --project <project> --zone <zone> --ssh-env gce --images <comma separated images>
+// $ godep go run run_e2e.go --logtostderr --v 2 --project <project> --zone <zone> --ssh-env gce --images <comma separated images>
 package main
 
 import (
@@ -42,6 +42,8 @@ var zone = flag.String("zone", "", "gce zone the hosts live in")
 var project = flag.String("project", "", "gce project the hosts live in")
 var images = flag.String("images", "", "images to test")
 var hosts = flag.String("hosts", "", "hosts to test")
+var cleanup = flag.Bool("cleanup", true, "If true remove files from remote hosts and delete temporary instances")
+var buildOnly = flag.Bool("build-only", false, "If true, build e2e_node_test.tar.gz and exit.")
 
 var computeService *compute.Service
 
@@ -53,6 +55,11 @@ type TestResult struct {
 
 func main() {
 	flag.Parse()
+	if *buildOnly {
+		// Build the archive and exit
+		e2e_node.CreateTestArchive()
+		return
+	}
 
 	if *hosts == "" && *images == "" {
 		glog.Fatalf("Must specify one of --images or --hosts flag.")
@@ -142,7 +149,7 @@ func main() {
 
 // Run tests in archive against host
 func testHost(host, archive string) *TestResult {
-	output, err := e2e_node.RunRemote(archive, host)
+	output, err := e2e_node.RunRemote(archive, host, *cleanup)
 	return &TestResult{
 		output: output,
 		err:    err,
@@ -154,7 +161,9 @@ func testHost(host, archive string) *TestResult {
 // Delete the instance afterward.
 func testImage(image, archive string) *TestResult {
 	host, err := createInstance(image)
-	defer deleteInstance(image)
+	if *cleanup {
+		defer deleteInstance(image)
+	}
 	if err != nil {
 		return &TestResult{
 			err: fmt.Errorf("Unable to create gce instance with running docker daemon for image %s.  %v", image, err),
